claw-and-order/requirements.txt
fastapi==0.135.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]>=2.0.29
asyncpg>=0.29.0
aiosqlite==0.19.0
discord.py==2.6.4
aiohttp>=3.9.0,<4.0
bcrypt==4.1.2
python-dotenv==1.0.1
slowapi==0.1.9
